This is a bit of code taken from Wiley Linux Bible
Processing shell script showing the capability of C code. (In this case I needed to see for myself I could use the Anjuntu to compile the code and then run it in the shell.) No changes proposed maybe more focus directed! (In this message just trying to get the Idea behind this forward momentum with GitHub and my inadequacies, relative to the systems level computer world as well as my abaility to capture forward momentum!!!!!!